Fix error with archive buttons
there was an error in the logic to prevent duplicate IDs where the
location could only be used once but it did not store the ID with the
location. This prevented multiple instances of the button on the
archive pages. This fix addresses that by storing the ID.

This also includes a version bump to 1.0.1 and necessary changes to the
readme.txt file for the version bump.